requirements.txt
python-telegram-bot
emoji
flask